src/pyserve/requirements.txt
Flask
flask_httpauth
simplejson
Flask-Cors
FUTIL
mysql-connector-python
